+++ packages/1ui/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './currency'
 export * from './general'
+export * from './status'
 export * from './transactions'
