+++ packages/1ui/src/types/status.ts
@@ -1,0 +1,8 @@
+export const QuestStatus = {
+  notStarted: 'not-started',
+  inProgress: 'in-progress',
+  claimable: 'claimable',
+  completed: 'completed',
+} as const
+
+export type QuestStatusType = (typeof QuestStatus)[keyof typeof QuestStatus]
