+++ packages/1ui/src/types/transactions.ts
@@ -13,14 +13,6 @@ export const TransactionStatus = {
 export type TransactionStatusType =
   (typeof TransactionStatus)[keyof typeof TransactionStatus]
 
-export const QuestStatus = {
-  notStarted: 'not-started',
-  inProgress: 'in-progress',
-  claimable: 'claimable',
-  completed: 'completed',
-} as const
-
-export type QuestStatusType = (typeof QuestStatus)[keyof typeof QuestStatus]
 export const Transaction = {
   identity: 'identity',
   claim: 'claim',
